requirements.txt
PyMuPDF
google-generativeai
streamlit
speechrecognition
pyaudio
textblob
whisper
torch
pydub